// src/lib/api/types/campaignSummaryResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Adisu Backend API
 * Backend API for the Adisu platform
 * OpenAPI spec version: 1.0.0
 */

export interface CampaignSummaryResponse {
  /** Campaign type */
  campaignType: string;
  /** Creation time */
  createdAt: string;
  /** Creator account ID */
  createdBy: string;
  /** Campaign ID */
  id: string;
  /** Campaign name */
  name: string;
  /** Scheduled sending time */
  scheduledFor?: string;
  /** Actual sending time */
  sentAt?: string;
  /** Campaign status */
  status: string;
}